Store UTC in PlistDate without DST conversions
To express more accuractly what is stored in the plist files, keep the
date within the UTC timezone, without DST conversion. If the isdst flag is
set to 0 or 1 it will roll the time back or forwards if the provided date
is subject to daylight saving within the local time zone.